REFACTORING #2 Architecture optimization No. 1
* created class MessageProcessor
* moved the message processing logic there

Issue: 2
